move randSpaceStrGener to builtin_string_vec_test file
