+++ expression/bench_test.go
@@ -373,29 +373,6 @@ func (g *randHexStrGener) gen() interface{} {
 	return string(buf)
 }
 
-type randSpaceStrGener struct {
-	lenBegin int
-	lenEnd   int
-}
-
-func (g *randSpaceStrGener) gen() interface{} {
-	n := rand.Intn(g.lenEnd-g.lenBegin) + g.lenBegin
-	buf := make([]byte, n)
-	for i := range buf {
-		x := rand.Intn(150)
-		if x < 10 {
-			buf[i] = byte('0' + x)
-		} else if x-10 < 26 {
-			buf[i] = byte('a' + x - 10)
-		} else if x < 62 {
-			buf[i] = byte('A' + x - 10 - 26)
-		} else {
-			buf[i] = byte(' ')
-		}
-	}
-	return string(buf)
-}
-
 // dataTimeStrGener is used to generate strings which are dataTime format
 type dataTimeStrGener struct{}
